+++ src/Version/SetDateForChangelogReleaseListener.php
@@ -7,7 +7,7 @@
 
 declare(strict_types=1);
 
-namespace Phly\KeepAChangelog\Changelog;
+namespace Phly\KeepAChangelog\Version;
 
 use Phly\KeepAChangelog\Common\ChangelogEditor;